feat(dracthyr): normalize Dracthyr race ID from 35 to 27 and resolve display ID collisions
- Stage 3: Normalization of Dracthyr race from ID 35 to ID 27, ensuring safe 32-bit raceMask boundaries. Added/updated DBC files.
- Stage 5: Added fallback Shaman skill line and armor/shield proficiency checks for Dracthyr Shamans in Player::LearnDefaultSkills().
- Stage 55: Deconflicted Dracthyr Dragon display IDs by implementing a range-based dynamic base offset system (Male: 1,100,000; Female: 1,101,000) in TwoForms.cpp.

+++ src/server/game/Entities/Player/Player.h
@@ -335,8 +335,8 @@ struct PlayerInfo
     float positionY{0.0f};
     float positionZ{0.0f};
     float orientation{0.0f};
-    uint16 displayId_m{0};
-    uint16 displayId_f{0};
+    uint32 displayId_m{0};
+    uint32 displayId_f{0};
     PlayerCreateInfoItems item;
     PlayerCreateInfoSpells customSpells;
     PlayerCreateInfoSpells  castSpells;
